Add files via upload
House Price Prediction Model

This repository contains a house price prediction model that is trained on a dataset of historical house sales. The model can be used to predict the price of a house based on its features, such as square footage, number of bedrooms and bathrooms, and location.

The model is implemented using the Python programming language and the scikit-learn machine learning library. The model is trained using a Linear Regression

The model can be used to predict the price of a house by providing it with the features of the house. The model will then output a prediction of the house's price.

The model can be used by homeowners, real estate agents, and investors to predict the price of a house. This information can be used to make informed decisions about buying, selling, or investing in real estate.

This model is for educational purposes only. It is not intended to be used as a professional tool for predicting house prices. The model's accuracy may vary depending on the quality of the training data and the features that are used to make predictions.